Rename /profile to /account
